+++ src/server/index.ts
@@ -1,5 +1,5 @@
 export {httpServer} from "./http";
 export {getModule} from "./bootstrap";
 export {Request} from "./controller-resolver";
-export {StatusCode} from "./status-code";
+export {Status} from "./status-code";
 export {fakeHttpServer, FakeServerApi, FakeResponseApi, IFakeServerConfig} from "./fake-http";